refactor: 이벤트 타입과 상수를 AppEvents.ts로 분리
