+++ src/pubsub/EventBus.ts
@@ -1,13 +1,4 @@
-import { MovieResponse } from "../../types/types";
-
-export type AppEventPayloads = {
-  loadStart: void;
-  moviesLoaded: MovieResponse;
-  searchLoaded: MovieResponse;
-  moreLoaded: MovieResponse;
-  titleChanged: string;
-  logoClick: void;
-};
+import { AppEventPayloads } from "./AppEvents";
 
 type Handler<T> = (payload: T) => void;
 
@@ -33,12 +24,3 @@ class EventBus {
 }
 
 export const eventBus = new EventBus();
-
-export const APP_EVENTS = {
-  LOAD_START: "loadStart",
-  MOVIES_LOADED: "moviesLoaded",
-  SEARCH_LOADED: "searchLoaded",
-  MORE_LOADED: "moreLoaded",
-  TITLE_CHANGED: "titleChanged",
-  LOGO_CLICK: "logoClick",
-} as const satisfies Record<string, keyof AppEventPayloads>;
